add support for latency command line argument
Added support for passing an optional microsecond latency
argument on the commandline. If no argument is specified,
it falls back to the default ~6 millisecond latency value.

This argument only affects the Pulseaudio sink buffer.
It does not account for the input buffer latency or
the intermediate buffer latency, both of which add
up to probably around 10 milliseconds.