refactor ring buffer math to utilize full buffer and fix PCM channel swap
The previous ring buffer implementation would always leave an empty
space in the buffer. This itself wasn't an issue, but it meant that
the buffer would always have an odd number of sample spaces available.

In the case of the PCM sink, samples are normally added in left+right
pairs. However, if the buffer was nearing max capacity, it was possible
to end up enqueueing only one of the samples of the pair, which would
then cause the left and right channels to be swapped forever.

This didn't affect the AC3 sink since there was an explicit check
to discard the complete 6 channel frame if the whole thing wouldn't
fit into the buffer.

With the new implementation, the complete buffer is used. Since the
buffer is a power of 2 in size and samples are added in pairs, this
issue is avoided and no explicit check is needed.